+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Removed deprecated method `Collection::saveDocument()`
   * Removed deprecated method `Client::setConnection()`. Use Client::setMongoClient().
   * Removed deprecated method `Client::getConnection()`. Use Client::getMongoClient().
+  * Removed `Document::pushFromArray()`. Use `Document::pushEach()`.
 
 ## 1.12.8 (2015-03-03)
   * Method 'Validator::validate()' marked as final